Nettoyage de code

+++ src/main/java/fr/iut/montreuil/stationski/Modele/Projectile.java
@@ -1,16 +1,15 @@
 package fr.iut.montreuil.stationski.Modele;
 
-import javafx.beans.property.SimpleIntegerProperty;
 
 public abstract class Projectile extends Entite{
     private Acteur cible;
-    private int ptsAttaque;
-    public Projectile(Acteur cible,  int posX, int posY, int ptsAttaque){
+    private double ptsAttaque;
+    public Projectile(Acteur cible,  int posX, int posY, double ptsAttaque){
         super(posX, posY);
         this.cible = cible;
         this.ptsAttaque = ptsAttaque;
     }
-    public int getPtsAttaque(){return this.ptsAttaque;}
+    public double getPtsAttaque(){return this.ptsAttaque;}
     public Acteur getCible() {
         return this.cible;
     }
